state-proof-based catchup
Add support for catchup to use state proofs to validate new blocks.

The main changes are:

- The catchup service now has a state proof fetcher, whose job is to
  retrieve state proofs for rounds beyond the current ledger state.
  These state proofs are stored in an sqlite DB, since there might be
  many state proofs that we need to fetch.

- The catchup service exposes a way to set trusted "renaissance"
  parameters for authenticating the initial state proofs, for cases
  when we can't use state proofs from the genesis block (like the
  situation we have on mainnet now).

- The BlockService HTTP interface adds support for retrieving a
  state proof, and for getting a light block header proof instead
  of a cert when retrieving a block.

- The catchup service uses state proofs, if possible, to authenticate
  new blocks, in lieu of agreement certificates.  The catchup service
  is backwards-compatible: if it requests a state proof from the
  BlockService, but receives an agreement certificate instead (e.g.,
  because the BlockService has not been upgraded with the above
  changes), the catchup service will validate the certificate instead.

- The config file has additional fields to optionally specify the
  renaissance catchup parameters, to allow catchup to start validating
  state proofs from some (trusted through out-of-band channels) block.

It may be a good idea, for performance, to pre-compute the state proofs
and distribute them in a single file, rather than asking many relays to
find the state proofs on-demand.  This is not done yet, because it largely
depends on how we would want to distribute these bundled state proofs.
It should be reasonably straightforward to feed a bundle of state proofs
into the stateProofFetcher.

+++ stateproof/stateproofMessageGenerator.go
@@ -59,7 +59,7 @@ func GenerateStateProofMessage(l BlockHeaderFetcher, round basics.Round) (statep
 	}
 
 	proto := config.Consensus[latestRoundHeader.CurrentProtocol]
-	votersRound := uint64(round.SubSaturate(basics.Round(proto.StateProofInterval)))
+	votersRound := round.SubSaturate(basics.Round(proto.StateProofInterval))
 	commitment, err := createHeaderCommitment(l, &proto, &latestRoundHeader)
 	if err != nil {
 		return stateproofmsg.Message{}, err
@@ -75,7 +75,7 @@ func GenerateStateProofMessage(l BlockHeaderFetcher, round basics.Round) (statep
 		VotersCommitment:       latestRoundHeader.StateProofTracking[protocol.StateProofBasic].StateProofVotersCommitment,
 		LnProvenWeight:         lnProvenWeight,
 		FirstAttestedRound:     votersRound + 1,
-		LastAttestedRound:      uint64(latestRoundHeader.Round),
+		LastAttestedRound:      latestRoundHeader.Round,
 	}, nil
 }
